requirements.txt
openpyxl==3.0.9
pandas==1.2.4
faiss-gpu==1.7.1.post2
transformers==4.12.0
tqdm==4.49.0
flask-restful==0.3.9